refact: AttractiveTypeCode enum 분리
